Migrating sign-in screen to auth module
