+++ androidApp/src/main/java/com/leandrodev/contasdolar/android/di/DependencyInjectionProvider.kt
@@ -1,7 +1,7 @@
 package com.leandrodev.contasdolar.android.di
 
 import com.leandrodev.bills.billsDiModule
-import com.leandrodev.contasdolar.android.auth.signin.authModule
+import com.leandrodev.auth.di.authModule
 import com.leandrodev.contasdolar.android.auth.state.authStateModule
 import com.leandrodev.bills.wallet.list.WalletListViewModel
 import com.leandrodev.bills.wallet.list.WalletListViewModelImpl
