--- androidApp/src/main/java/com/leandrodev/contasdolar/android/model/User.kt
@@ -1,5 +1,0 @@
-package com.leandrodev.contasdolar.android.model
-
-
-class User {
-}
