--- androidApp/src/main/java/com/leandrodev/contasdolar/android/utils/FirebaseUtils.kt
@@ -1,8 +1,0 @@
-package com.leandrodev.contasdolar.android.utils
-
-import com.leandrodev.contasdolar.android.model.User
-import dev.gitlive.firebase.auth.FirebaseUser
-
-fun FirebaseUser.toUser(): User {
-    return User()
-}
