+++ androidApp/src/test/kotlin/unit/auth/AuthViewModelImplTest.kt
@@ -1,12 +1,12 @@
 package unit.auth
 
-import com.leandrodev.contasdolar.android.auth.signin.AuthViewModelImpl
+import com.leandrodev.auth.signin.SignInViewModelImpl
 import org.junit.Test
 
 class AuthViewModelImplTest {
     @Test
     fun `given initial state, when click on google login, invoke google login business`() {
-        val viewModel = AuthViewModelImpl()
+        val viewModel = SignInViewModelImpl()
         viewModel.onGoogleAuthClick()
 
         // TODO Verify ViewModel dependency
@@ -14,7 +14,7 @@ class AuthViewModelImplTest {
 
     @Test
     fun `given initial state, when click on google login, should update to loading state`() {
-        val viewModel = AuthViewModelImpl()
+        val viewModel = SignInViewModelImpl()
         viewModel.onGoogleAuthClick()
 
         // TODO Verify observers on ViewModel
