+++ auth/build.gradle.kts
@@ -8,3 +8,14 @@ android {
         jvmTarget = "11"
     }
 }
+
+kotlin {
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                implementation(project(":shared"))
+                implementation(project(":ui"))
+            }
+        }
+    }
+}
