+++ auth/src/androidMain/kotlin/com/leandrodev/auth/di/AuthModuleAndroid.kt
@@ -1,0 +1,10 @@
+package com.leandrodev.auth.di
+
+import com.leandrodev.auth.signin.SignInProvider
+import com.leandrodev.auth.signin.SignInProviderImpl
+import org.kodein.di.DI
+import org.kodein.di.bindSingleton
+
+actual val authPlatformModule: DI.Module = DI.Module("AuthAndroid") {
+    bindSingleton<SignInProvider> { SignInProviderImpl() }
+}
