+++ auth/src/androidMain/kotlin/com/leandrodev/auth/utils/FirebaseUtils.kt
@@ -1,0 +1,8 @@
+package com.leandrodev.auth.utils
+
+import com.leandrodev.auth.model.User
+import dev.gitlive.firebase.auth.FirebaseUser
+
+fun FirebaseUser.toUser(): User {
+    return User()
+}
