+++ auth/src/androidMain/res/values/lazy_values.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <item name="default_web_client_id" type="string" />
+</resources>
