+++ auth/src/commonMain/kotlin/com/leandrodev/auth/model/User.kt
@@ -1,0 +1,4 @@
+package com.leandrodev.auth.model
+
+class User {
+}
