+++ auth/src/iosMain/kotlin/com/leandrodev/auth/di/AuthModuleIos.kt
@@ -1,0 +1,5 @@
+package com.leandrodev.auth.di
+
+import org.kodein.di.DI
+
+actual val authPlatformModule: DI.Module = DI.Module("AuthIos") {}
